Make the SE driver provider-agnostic
The SE Driver was only targeting the TPM provider. This behaviour has
been changed to be provider-agnostic: by default the SE driver will
target the highest priority provider returned by the ListProviders
operation. Features are introduced to also have a way to select a single
provider.

This commit also pull the newest psa-crypto with the new `interface`
feature. This SE driver does not need to link with the PSA Crypto
implementation but only needs the interface implementation.

Fix #7

Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>